Bug 1198458: Rollup of changes previously applied to media/webrtc/trunk/webrtc and fixes to those rs=jesup r=froyd,jib,bwc,jesup,gcp,sotaro,pkerr,pehrsons Landing as one rolled-up patch to avoid breaking regression tests, and in keeping with previous WebRTC imports.  Broken out parts that needed review are on the bug.
UltraBlame original commit: ae32ad44ce1c048258a881dbaa7727def08c53a4

+++ dom/media/systemservices/CamerasChild.h
@@ -85,7 +85,7 @@ public:
   
   
   virtual bool RecvDeliverFrame(const int&, const int&, mozilla::ipc::Shmem&&,
-                                const int&, const uint32_t&, const int64_t&,
+                                const size_t&, const uint32_t&, const int64_t&,
                                 const int64_t&) override;
   virtual bool RecvFrameSizeChange(const int&, const int&,
                                    const int& w, const int& h) override;
